+++ examples/index_resources.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 ###
-#  (C) Copyright (2012-2017) Hewlett Packard Enterprise Development LP
+# (C) Copyright (2012-2017) Hewlett Packard Enterprise Development LP
 #
 # Permission is hereby granted, free of charge, to any person obtaining a copy
 # of this software and associated documentation files (the "Software"), to deal
@@ -20,19 +20,36 @@
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 # THE SOFTWARE.
 ###
+from pprint import pprint
 
+from hpOneView.oneview_client import OneViewClient
+from config_loader import try_load_from_file
 
-from setuptools import find_packages
-from setuptools import setup
+config = {
+    "ip": "<oneview_ip>",
+    "credentials": {
+        "userName": "<username>",
+        "password": "<password>"
+    }
+}
 
-setup(name='hpOneView',
-      version='4.2.0',
-      description='HPE OneView Python Library',
-      url='https://github.com/HewlettPackard/python-hpOneView',
-      download_url="https://github.com/HewlettPackard/python-hpOneView/tarball/v4.2.0",
-      author='Hewlett Packard Enterprise Development LP',
-      author_email='oneview-pythonsdk@hpe.com',
-      license='MIT',
-      packages=find_packages(exclude=['examples*', 'tests*']),
-      keywords=['oneview', 'hpe'],
-      install_requires=['future>=0.15.2'])
+attribute = 'Model'
+category = 'server-hardware'
+
+# Try load config from a file (if there is a config file)
+config = try_load_from_file(config)
+
+oneview_client = OneViewClient(config)
+
+print('\nGetting all index resources:')
+index_resources = oneview_client.index_resources.get_all()
+pprint(index_resources)
+
+sh = oneview_client.server_hardware.get_all()[0]
+print('\nGetting index resource for server hardware with uri "{0}":'.format(sh['uri']))
+index_resource = oneview_client.index_resources.get(sh['uri'])
+pprint(index_resource)
+
+print('\nGetting aggregated index resources with attribute: "{0}" and category: "{1}"'.format(attribute, category))
+index_resources = oneview_client.index_resources.get_aggregated(attribute, category)
+pprint(index_resources)